Use `muladd` explicitly inside `contract`.
This avoids a need to call `mulexpr`. The `@turbo` preprocessing
replaces `a + b*c` (and similar expressions, like `a += b*c`) with
`muladd`, in order to simplify some downstream code (i.e. cost
modelling). However, due to what is probably a Julia deserialization
bug on Julia 1.8.5, this can cause problems when building
sysimages. Thus, we avoid the need for doing it manually here via just
calling `muladd` in the first place, rather than requiring the
substitution to happen later.
Plus, not requiring it should be faster by an extremely minor amount.